feat: InterviewCamera 내부 state 정의
